[ALPS03192729] linux kernel patch: CVE-2017-6074: device specific
[Detail]
In the current DCCP implementation an skb for a DCCP_PKT_REQUEST packet
is forcibly freed via __kfree_skb in dccp_rcv_state_process if
dccp_v6_conn_request successfully returns.

However, if IPV6_RECVPKTINFO is set on a socket, the address of the skb
is saved to ireq->pktopts and the ref count for skb is incremented in
dccp_v6_conn_request, so skb is still in use. Nevertheless, it gets freed

[Solution]
However, if IPV6_RECVPKTINFO is set on a socket, the address of the skb
is saved to ireq->pktopts and the ref count for skb is incremented in
dccp_v6_conn_request, so skb is still in use. Nevertheless, it gets freed
in dccp_rcv_state_process.

Fix by calling consume_skb instead of doing goto discard and therefore
calling __kfree_skb.
Similar fixes for TCP:
fb7e239 [TCP]: skb is unexpectedly freed.
0aea76d tcp: SYN packets are now
simply consumed

 Changes to be committed:
        modified:   net/dccp/input.c

Feature:others
Signed-off-by: Andrey Konovalov <andreyknvl@google.com>
Acked-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: mtk14233 <Frank.Yu@mediatek.com>
Change-Id: Ie4ce67a629e8069e75aa0a87d8a3f8202d0d0ecd
CR-Id: ALPS03192729
(cherry picked from commit 9e150113cbe2f7bbc602f209cabcd421024f5811)