[ALPS02613621] soc: mediatek: pwrap:remove pwrap_is_mt8135/pwrap_is_mt8173
With more SoCs being added the list of helper functions like these would
grow. To mitigate this problem we remove the existing helpers and change
the code to test against the pmic type stored inside the pmic specific
datastructure that our context structure points at. There is one usage of
pwrap_is_mt8135() that is ambiguous as the test should not be dependent on
mt8135, but rather on the existence of a bridge. Add a new element to
pmic_wrapper_type to indicate if a bridge is present and use this where
appropriate.

Signed-off-by: John Crispin <blogic@openwrt.org>
Signed-off-by: Matthias Brugger <matthias.bgg@gmail.com>
(cherry picked from commit 25269cefb691d324a3fe2be4d2a7b81f6a733996)

Change-Id: Ic8d83c1a497f22805e328f9d34168e49828a1b11
Signed-off-by: Chen Zhong <chen.zhong@mediatek.com>
CR-Id: ALPS02613621
Feature: PMIC_WRAP