[ALPS02677544] cmo: cache maintenance operations for MediaTek
[Detail]
Interfaces for cache maintenance operations, which includes:

1) inner_dcache_flush_all() to flush the L1 dcache of the
caller and flush L2 dcache of the cluster that the caller
located in.

2) smp_inner_dcache_flush_all() to flush the L1 dcaches of
all the online cores, and flush the L2 dcaches of all the
online clusters. This function leverages smp call to flush
dcaches concurrently, thus it CANNOT be used in interrupt
context.

Change-Id: I906172de1bf2d6cc6fe7e1477dc8b2a08dc74456
Signed-off-by: Shun-Chih Yu <shun-chih.yu@mediatek.com>
CR-Id: ALPS02677544
Feature: Others