[ALPS02734453] partition: refine sysenv
[Detail]
sysenv is a feature to store data which can be accessed by
both kernel and lk. This module was a kernel module with
raw partition read and write behavior. For the requirement of
kernel standardization, raw read and raw write partitions in
kernel driver are forbidden. So move this module to userspace.

[Solution]
Move sysenv to userspace, and keep kernel module for backward
compatible. The next stage will remove related kernel module
and all users have to change to use sysenv userspace API.

Change-Id: If84e7ef36c9fe5c9e5127a85f3c7fa73069f4747
Signed-off-by: Chang-An Chen <chang-an.chen@mediatek.com>
CR-Id: ALPS02734453
Feature: eMMC Boot Up