[ALPS02595342] sched: support sched-assist dvfs
[Detail]
The sched-assist is NOT a governor of DVFS, which provide
hints for governor(interactive) to change frequency
That make governor have a view of task level in order to get
better performance and power. This feature needs to be dependent
on CONFIG_CPU_FREQ_GOV_SCHED.

+CONFIG_CPU_FREQ_SCHED_ASSIST

To fixup conflict between schedDVFS and hotplug also.

[Solution]
Add hint of scheduler for interactive DVFS and hotplug.

Change-Id: Ie72ff1c4cb5842419ada4a74179f19cba99eed2d
CR-Id: ALPS02595342
Signed-off-by: Michael Jiang <michael.jiang@mediatek.com>
Feature: performance and power balance customization