[ALPS02875181] wlan: Bug fix of repeating oid complete
[Detail]
The wlanOidSetCountryCode is the oid handler of kalIoctl,
if there is any oid cmd to send in wlanOidSetCountryCode,
the return value of wlanOidSetCountryCode must be WLAN_STATUS_PENDING.
In rlmDomainSendCmd, the second parameter is TRUE,
means the cmd we send is oid.
We found two oid cmd, so we complete oid repeatly.

[Solution]
Modify the second parameter of rlmDomainSendCmd to FALSE.

Change-Id: I8b742bdd2306456fe1c191b215378e96cb7e12a5
Signed-off-by: ziming sun <ziming.sun@mediatek.com>
CR-Id: ALPS02875181
Feature: [Android Default] Wi-Fi Driver for SoC