[ALPS02978361] PSEUDO M4U: Do not call find_vma for kernel thread
[Detail]
If we call find_vma for kernel thread which do not have current->mm,
we may got the kernel panic as following:
Unable to handle kernel paging request at virtual address at xxxxx,
PC is at find_vma+0x40/0x88 LR is at find_vma+0x1c/0x88.

[Solution]
For kernel thread which use active_mm, it's not supposed to access the
userspace va, and we may not get the vma for this current of thread.
Even it exist, we may get the wrong vma.
Do not call find_vma for kernel thread.

Change-Id: Idfa095004b4a6edb5766e3521bbbf164f339e9d5
Signed-off-by: Honghui Zhang <honghui.zhang@mediatek.com>
CR-Id: ALPS02978361
Feature: PSEUDO M4U