[ALPS02997552] Pseudo m4u: handle mm fault if could not get the pa
[Detail]
When userspace alloca mva with VM_PFNMAP flag set, we should get
the pa from system pagetable, and we may could not get the pa if
the page is swap out, we need to handle mm fault to fore the page
swap in.

[Solution]
Handle mm fault to force the page swap in if the page was swap out.

Change-Id: I62429cf853ac3caf3f24538a7553d2b7365b1fa0
Signed-off-by: Honghui Zhang <honghui.zhang@mediatek.com>
CR-Id: ALPS02997552
Feature: Pseudo m4u