[ALPS02864812] timer: fix hang_detected flag for timer interrupt
[Detail] Fix Linux kernel bug: hang_detected shall be reset while
CPU plug-on to ensure future timer interrupt works fine in such
cpu.

Details:
If hang_detected is 1 in this new CPU, after the tick device binding
to this CPU is switched to HRTimer, this CPU will NOT do
tick_program_event() for its tick device because hang_detected is 1.
In the end, this CPU will NOT have any tick event in the future.

To avoid above case, we shall reset it specifically while CPU is
plugged-on or CPU is plugged-off.

Change-Id: If94fa667aa013eea8efdf641b0713abd71365b78
Signed-off-by: Stanley Chu <stanley.chu@mediatek.com>
CR-Id: ALPS02864812
Feature: Others