[ALPS03052379] Wi-Fi: TGac 5.2.63 fail to pass STS capability check
[Detail]
For normal user scenario, we will set STS number according to AP's
dimensions number to avoid IOT issue.
In 5.2.63's test, testbed AP's dimensions number is set to 2, so we set
STS number to 2.
But this will fail 5.2.63 step3 where beamformee STS number
is expected to be the maximum number of beamformer antennas we can
support.

[Solution]
Add "ForceSTSNum" setting in wifi.cfg and set it to 1 for MT6632 to pass
TGac cert. test.

Change-Id: I55ea3fcd904dee781649910ce58db518fb6c11a1
CR-Id: ALPS03052379
Feature: Wi-Fi
Signed-off-by: Chaozhong.Liang <chaozhong.liang@mediatek.com>