[ALPS03070988] Timer: apply patch "Plug locking race vs. timer migration"
[Detail] Apply Linux kernel patch:
Plug locking race vs. timer migration

Linus noticed that lock_timer_base() lacks a READ_ONCE() for accessing the
timer flags. As a consequence the compiler is allowed to reload the flags
between the initial check for TIMER_MIGRATION and the following timer base
computation and the spin lock of the base.

While this has not been observed (yet), we need to make sure that it never
happens.

Change-Id: I4545f5d1e56994cf8375dfe943e18eb7d65d8033
Signed-off-by: Stanley Chu <stanley.chu@mediatek.com>
CR-Id: ALPS03070988
Feature: CTS(*)