[ALPS03123639] Sound: No Ringtone After Play FM
[Detail]
Connect bt heedset and earphone, and play fm.
when MT call coming, can't hear the ringtone.

[Solution]
The Fm pcm driver not successfully close the afe.
So the apll is still enable. but someone close it.
but the afe flag is still true, so the next sound
will not open the apll.

Fix the fm driver close sequence, so the next sound
will open apll normally.

Change-Id: I6412f693756b8aa1393347aaa5155dc44441fc30
Signed-off-by: chien-wei hsu <chien-wei.hsu@mediatek.com>
CR-Id: ALPS03123639
Feature: Call Control Framework