[ALPS02796449] mm: detect high-end page allocation failures
To test if our drivers work fine with large memory system,
we need our drivers to use high-end (>=4GB) memory as much
as possible.

It is possible that the high-end memory runs out and the driver
get a fall-back allocated non-high-end memory (<4GB).
The system still works normal in fall-back cases.
However our goal is to test drivers with high-end memory, so we
add an warning if the fall-back allocation happens.

Change-Id: I0d51bd0e807926c594f350ba46ff82dc9dc28f02
CR-Id: ALPS02796449
Feature: Large DRAM
MTK-Commit-Id: 32d5816a6b708e4c6c5449ba317e7249e1f5a34f