[ALPS02613621] soc: mtk-pmic-wrap: avoid integer overflow warning
On ARM64, the mtk-pmic-wrap driver causes a harmless warning:

mtk-pmic-wrap.c:1062:16: warning: large integer implicitly
	truncated to unsigned type [-Woverflow]
mtk-pmic-wrap.c:1074:16: warning: large integer implicitly
	truncated to unsigned type [-Woverflow]
mtk-pmic-wrap.c:1086:16: warning: large integer implicitly
	truncated to unsigned type [-Woverflow]
  .int_en_all = ~(BIT(31) | BIT(1)),

The problem is that the result of the BIT() macro is an unsigned long,
so taking the bitwise NOT operation of that results in an integer
with the upper 32 bits all set and that cannot be assigned to a
u32 variable without loss of information.

This is harmless because we were never interested in the upper bits
here anyway, so we can shut up the warning by adding a simple cast
to u32.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
(cherry picked from commit 7f058c0a4a0441044f96fcc863533004bc82cf65)

Change-Id: Ic8cd553805ac4c8f77bc6bd6601c56ec81805ed6
CR-Id: ALPS02613621
Feature: PMIC_WRAP
MTK-Commit-Id: 70aae2ab0e362b9d824953b979a90623c02156c1