[ALPS02766124] ftl: fix recover block error issue
[Detail]
For the first boot, there will go to do recover block action.
But it just use the first 4 bytes to judge whether the pmt page
is valid. This is not safe, because the sector number may be not
continully. Then, part of pmt may be lost when boot. This can
result to the mount fs fail.

[Solution]
Check the whole page whether is all 00 or FF.

Change-Id: I24e39108623f7fa812a348a1d61b2399f5c28867
CR-Id: ALPS02766124
Feature: TLC NAND boot up
MTK-Commit-Id: ebcb298ac6c64dfadccf860740751376b3f71bf1