[ALPS02924520] mem: switch to lowpower when init
Switch the DRAM setting to lowpower mode in DCS init function.

Previous implementation assumes the DRAM can be in lowpower
mode even the kernel DCS driver is disabled. So we need addition
APIs to handle this situation and it leads to a bad design.

This patch assumes that the DRAM is in normal mode before
its initialization. The DCS H/W should be disabled when the
DCS kernel driver is off. The assumption works with platforms
which do not have the DCS H/W.

Change-Id: I03a9861bac309a85796efe70dd6845ae56d130e0
CR-Id: ALPS02924520
Feature: DCS (Dynamic Channel Switch)
MTK-Commit-Id: 87894da20ae0ab4063fffc73525913774df8b1c7