[ALPS02949244] AEE: revise kernel API dump flow
[Detail] there are two processes may access the global variable of aed
driver, one is debuggerd and the other is kworker. The working flow that
we expected is as following:
1. kworker generates one oops one time and wait for child debuggerd to
handle it.
2. child debuggerd get the oops and output one DB as a result, then
notify kworker by complete() function.
3. kworker clean up the related message of this oops.
4.kworker check if there is other oops need to be generated, and repeat
step 1.2.3.4 if yes.

However, debuggerd may have chance to do step 2 again while kworker was
still doing setp 3 or the setp 1 of next oops, which lead to serious
problem.

Example 1, there is chance that debuggerd fork child process to handle
one oops during the time between oops is passed to .lastlog and the .msg
successfully being assigned. Refer to CR ALPS02820034.

Example 2, while debuggerd is in the complete() function the kworker may
calling init_completion, which free the lock that complete() still
holding. Refer to CR alps02879469.

[Solution] add a variable to control debuggerd to fork child debuggerd
when it is safe.

Change-Id: Iff22d9a1fff27d72f1037171eb02d6111c79ef06
CR-Id: ALPS02949244
Feature: Android Exception Engine(AEE)
MTK-Commit-Id: 1c98418d416861f8d642933635f84ca0472ff552