[ALPS02961914] CPU: Errata 845719 patch
[Detail]
When executing in AArch32 state at EL0, a load to the same offset
withina different 4GB region as a recent previous load in AArch64 state
might read incorrect data.

[Solution]
The erratum can be avoided by causing one of the conditions given in
the list above to occur between positions 2 and 3 in the sequence.
Because there must be an exception return from AArch64 to AArch32
between positions 2 and 3 in the above sequence, the recommended
workaround is to insert a write of the CONTEXTIDR_EL1 register into
operating system exception return sequences.

Change-Id: I34c3914c31256eee547a4d5ebade8763bf464382
CR-Id: ALPS02961914
Feature: others
MTK-Commit-Id: f6ce07e6770b5ec74f3fb252b0cbfe8a090cfb7d