[ALPS02839895] sched: a configurable interface for capacity margin
[Detail]
The capacity margin is the margin of free capacity and essential
to frequecy change and task placement. A different margin for
different category might be needed, so add a configurable interface
for it. The value should be above 1024.
(default 1280, 20% free capacity)

echo 1280 > /sys/devices/system/cpu/eas/capacity_margin

[Solution]
a configurable interface for capacity margin.

Change-Id: Ia37df795e93f9891747f3acd47ab8284f16af17e
CR-Id: ALPS02839895
Feature: performance and power balance customization
MTK-Commit-Id: 03bfe6be73c054a7efa0abc79899f037cdf6f47f