[ALPS02978134] usb: core: fix a double free in the usb driver
[Detail]
There is a double free problem in the usb driver.
This is caused by delayed deregister for scsi device.

[Solution]
Make sure to deallocate the bandwidth_mutex only when the last HCD is
freed.  When hcd_release() is called for either hcd in a peer set,
invalidate the peers ->shared_hcd and ->primary_hcd pointers.

Change-Id: I16b4c8e4ffdc899850776e3b510d0f8643495a00
CR-Id: ALPS02978134
Feature: Common Framework
MTK-Commit-Id: 094d638d95a3a3dae53298b3e4b8caed985c8c16