[ALPS03015112] usb: Use CC AUXADC to detect cable in kpoc
If check VBUS when cable plugs out, the VBUS is decreasing
slowly. So the system will be turned off after several seconds.
For increasing the performance, the typec driver also detect
the voltage on CC pin. If CC is lower than cRd-Connect, just
notify the cable is out. But MUST know this way is OUT OF
type-c spec. So be careful.

Change-Id: I1ca52c910b39f498a7d46cd9df0a8de19b518f56
CR-Id: ALPS03015112
Feature: pump express charging. (fast charging)
MTK-Commit-Id: f54cfd781b160db450c1e4a9c79c8389a30ab6e6