[ALPS02773746] Sound: Dctrim Refactored
[Detail]
Remove Dctrim of AP side control.
New Dctrim is implemented in Pmic side.

The Mixer control of Dctrim is update
DcTrim_DC_Offset -> get/set the dc offset value
Dctrim_Control_Switch -> Control Dctrim of three state
"Not Yet", "Calibrated" show the status of Dctrim finished or not.
if set "Calibrating" to this, it will trigger the hardware to do
the dc calibration and update the dc_offset and hp_trimcode, the flag
also change to "Calibrated".

Change-Id: Ica9b6a611a543c6f4ffc7295d52a3d8e043f92fc
Signed-off-by: chien-wei hsu <chien-wei.hsu@mediatek.com>
CR-Id: ALPS02773746
Feature: Driver