[ALPS03209214] usb:unmap reqs in musb_gadget_queue()'s error case
If the descriptor is missing the request is never unmapped. This patch
changes this and renames the cleanup label to unlock since there is no
cleanup done. The cleanup would revert the allocation of resource (i.e.
this dma mapping) but it does not, it simply unlocks and returns.

Change-Id: I2064a2ac19f279126ea78064a54deabab3b69028
Signed-off-by: Justin.Hsieh <justin.hsieh@mediatek.com>
CR-Id: ALPS03209214
Feature: Others