[ALPS03254794] charger: return dummy temp when no VBUS
[Detail]
If USB cable is not plugged in and is not charging, return
dummy value to thermal driver for charging IC's temperature.
This will not trigger ADC measurement for charging IC and
can lower power consumption.
Only measure and return charging IC's temperature when USB
cable is plugged in.

Change-Id: I867c2b21d2e698e62d666f7099f0c50c7f298d31
Signed-off-by: Elvin Lin <elvin.lin@mediatek.com>
CR-Id: ALPS03254794
Feature: Charger