[ALPS03249427] dm-crypt: Fix incorrect storage type detection
[Detail] Fix incorret storage type detection in dm-crypt for hw fde:

In the past, we use major and minor number to decide the type of
encryptable storage device. This may lead to incorrect result because
major number varies according to number of partitions in a project.

Now we changed the way: Searching keyword in block device path.

For embedded storage, there shall be "bootdevice".
For external storage, there shall be "externdevice" or "vold".

Change-Id: Ia716f23bda3d684006fd5a044659340e4fd018b0
Signed-off-by: Stanley Chu <stanley.chu@mediatek.com>
CR-Id: ALPS03249427
Feature: [Android Default] MTP/PTP (Media File Transfer / Picture File Transfer)