[ALPS03423801] ANDROID: mnt: Fix next_descendent
next_descendent did not properly handle the case
where the initial mount had no slaves. In this case,
we would look for the next slave, but since don't
have a master, the check for wrapping around to the
start of the list will always fail. Instead, we check
for this case, and ensure that we end the iteration
when we come back to the root.

Change-Id: I56389732b65eeb73e6ab4eaf0b9c26fe52a15d32
Signed-off-by: Daniel Rosenberg <drosen@google.com>
Bug: 62094374
Signed-off-by: LF LIu <lf.liu@mediatek.com>
CR-Id: ALPS03423801
Feature: SP META Tool