pwm: add pwm-mediatek [1/2]
This patch is support for the PWM core found on current ARM base SoCs
made by MediaTek. This IP core supports 3 channels and has 2 operational
modes. There is the old mode, which is a classical PWM and the new mode
which allows the user to define bitmasks that get clocked out on the pins.
As the subsystem currently only supports PWM cores with the "old" mode,
we can safely ignore the "new" mode for now.

Test: OK

Change-Id: Ie1e841d220b9ded24cd7bdfc1df8736cab977aff
Signed-off-by: Zhi Mao <zhi.mao@mediatek.com>