[ALPS03379109] wlan: p2p enhance connection
[Detail]
This is a timing issue when wifi driver indicated the rx EAP PKT and
RX deauth to supplicant at the same time, but supplicant ingored the rx
EAP PKT and directly disconnect. But supplciant don't try to re-connect.

[Solution]
1.P2P GO will delay the deauth frame when dirver sent the EAP of failure
(code = 4).
2.[ALPS03397844]fix the timing issue when GO del_station and GC re-auth
request are a race condition, driver will ignore the delay event when
driver received the GO's deauth.

Change-Id: I1ff963c6826f2597c8da921072e0db0aa660dc89
Signed-off-by: Leo Kuo <leo.kuo@mediatek.com>
CR-Id: ALPS03379109
Feature: Wi-Fi Direct