[ALPS03292700] ASoC: deep buffer music seek has pop noise
[Detail]
music seek has pop noise in deep buffer playback.

[Solution]
The original solution is to extend the buffer size of
deep buffer kernel size. The cons of this workaround is
increasing the cost of HW SRAM. Another option is to
clear the kernel buffer in ack callback function,
after ALSA copy.

Change-Id: I2a08c47799de26f6364b6232d1b467369b20fb79
Signed-off-by: Shane Chien <Shane.Chien@mediatek.com>
CR-Id: ALPS03292700
Feature: Phone Sound