[ALPS03535873] sched: rt scheduler sets capacity requirement
[Detail]
RT tasks don't provide any running constraints like deadline ones
except their running priority. The only current usable input to
estimate the capacity needed by RT tasks is the rt_avg metric. We use
it to estimate the CPU capacity needed for the RT scheduler class.

In order to monitor the evolution for RT task load, we must
peridiocally check it during the tick.

Then, we use the estimated capacity of the last activity to estimate
the next one which can not be that accurate but is a good starting
point without any impact on the wake up path of RT tasks.

[Solution]
commit 6e1e1ed87206 ("sched: rt scheduler sets capacity requirement")

Change-Id: I0fcadb3db1c07aed6b99f7a92f21c670666165ea
CR-Id:ALPS03535873
Feature:Sched
Signed-off-by: Jing-Ting Wu <jing-ting.wu@mediatek.com>