[ALPS03569316] sched/tune: update accouting before CPU capacity
[Detail]
FIXUP: sched/tune: update accouting before CPU capacity

The SchedTune tasks accounting is used to identify how many tasks are in
a boostgroup and thus to bias the selection of an OPP based on the
maximum boost value of the active boostgroups.
The current implementation however update the accounting after CPU
capacity has been update. This has two effects:
a) when we enqueue a boosted task, we do not immediately boost its CPU
b) when we dequeue a boosted task, we can keep a CPU boosted even if not
   required

This patch change the order of the SchedTune accounting and SchedFreq
updated to ensure to have always an updated representation of which
boosted tasks are runnable on a CPU before updating its capacity.

Reported-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

[Solution]
commit a85045c03470
("FIXUP: sched/tune: update accouting before CPU capacity")

Change-Id: Ie170a4c96baf893d9f2cae960c8430abe1dec5ff
CR-Id:ALPS03569316
Feature:Sched
Signed-off-by: Jing-Ting Wu <jing-ting.wu@mediatek.com>