[ALPS03610692] ALSA: CVE-2017-13167
[Detail]
[PATCH] UPSTREAM: ALSA: timer: Call notifier in the same spinlock

snd_timer_notify1() is called outside the spinlock and it retakes the
lock after the unlock.  This is rather racy, and it's safer to move
snd_timer_notify() call inside the main spinlock.

The patch also contains a slight refactoring / cleanup of the code.
Now all start/stop/continue/pause look more symmetric and a bit better
readable.

Bug: 37240993

Change-Id: I5097827a363d70348122343dc26615cb5c045890
Signed-off-by: Michael Hsiao <michael.hsiao@mediatek.com>
CR-Id: ALPS03610692
Feature: Others