DO NOT MERGE: Modify mtk-pmic-keys to support gpio sysfs
The PMU_FCHR_ENB interrupt line is connected to a PMIC register, but
doubles as a volume down pin on some products. However, in certain
use-cases, we want to suppress these keyevents and instead directly read
the line state (high/low) from user-space. In order to support this, we
add functionality here to expose this interrupt as a virtual GPIO.

The feature is guarded by config KEYBOARD_MTK_PMIC_GPIO and only takes
effect if the device-tree sets the `mediatek,homekey-gpio` entry. A
virtual gpiochip is created with the provided base, with a single GPIO
that reflects the current interrupt state. This GPIO should be treated
as a read-only pin, and will reject any attempts to configure itself as
an output pin.

In this change, we rely on passing a simulated interrupt line back to
the GPIO subsystem to support edge-based triggers. The logic for this
line has been backported from mainline Linux's `irq_sim` files.

Bug: 118189998
Test: Verified using a test application that volume down key events on
both edges can be registered by applications.
Change-Id: I201e757d2ce20c64dd39b1c0819075244adc9d33