[ALPS05435417] Upstream: audit: fix error handling in audit_data_to_entry()
This security patch involved 2 related patches.

commit 70c4cf17e445 ("audit: fix a memory leak bug")
commit 2ad3e17ebf94 ("audit: fix error handling in audit_data_to_entry()")

Commit 219ca39427bf ("audit: use union for audit_field values since
they are mutually exclusive") combined a number of separate fields in
the audit_field struct into a single union.  Generally this worked
just fine because they are generally mutually exclusive.
Unfortunately in audit_data_to_entry() the overlap can be a problem
when a specific error case is triggered that causes the error path
code to attempt to cleanup an audit_field struct and the cleanup
involves attempting to free a stored LSM string (the lsm_str field).
Currently the code always has a non-NULL value in the
audit_field.lsm_str field as the top of the for-loop transfers a
value into audit_field.val (both .lsm_str and .val are part of the
same union); if audit_data_to_entry() fails and the audit_field
struct is specified to contain a LSM string, but the
audit_field.lsm_str has not yet been properly set, the error handling
code will attempt to free the bogus audit_field.lsm_str value that
was set with audit_field.val at the top of the for-loop.

This patch corrects this by ensuring that the audit_field.val is only
set when needed (it is cleared when the audit_field struct is
allocated with kcalloc()).  It also corrects a few other issues to
ensure that in case of error the proper error code is returned.

audit: fix a memory leak bug
Signed-off-by: Wenwen Wang <wang6495@umn.edu>
Reviewed-by: Richard Guy Briggs <rgb@redhat.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>
audit: fix error handling in audit_data_to_entry()
Cc: stable@vger.kernel.org
Fixes: 219ca39427bf ("audit: use union for audit_field values since they are mutually exclusive")
Signed-off-by: Paul Moore <paul@paul-moore.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Ivan Tseng <ivan.tseng@mediatek.com>
CR-Id: ALPS05435417
Feature: Google Security Patch
Change-Id: If2ea51b4a0849b9fe9b9b1929ed3095b82514244
(cherry-pick from 031347ddb432d5fb6d65949e966e5e8e977ad215)