Fix the SCM version of the repository
The kernel's version is based off the output of the setlocalversion
script, which takes into account CONFIG_LOCALVERSION_AUTO (which we do
not set) and LOCALVERSION (which we also do not set). Based on these
empty defaults, the script relies on the git history to determine
whether or not to append a '+' symbol to indicate that this is a kernel
with patches (e.g. '4.4.95+').

We rely on this '+' symbol as other proprietary code hard-codes these
paths. However, buildbots make no guarantees about the state of the .git
history (b/138144743#comment6), leading to scenarios where runs
conducted on different buildbots lead to the kernel being built with or
without the '+'. While this doesn't break the kernel, it does break
other binaries relying on the /sys/module/kernel path.

Separately, we should look into removing this dependency. For the short
term, we short-circuit this path by creating an .scmversion file. This
identifies the kernel as having been built with modified sources.

Bug: 138144743
Test: Verified that repeated builds result in the proper version string.
Change-Id: I997b705a45a3c3c5c255c9b58da346e83d65a36e