[ALPS05024007] HID: Fix assumption that devices have inputs
The syzbot fuzzer found a slab-out-of-bounds write bug in the hid-gaff
driver.  The problem is caused by the driver's assumption that the
device must have an input report.  While this will be true for all
normal HID input devices, a suitably malicious device can violate the
assumption.

The same assumption is present in over a dozen other HID drivers.
This patch fixes them by checking that the list of hid_inputs for the
hid_device is nonempty before allowing it to be used.

Change-Id: Id7619f6e05f8164f1dffa7779c34f75678cbf508
Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
CC: <stable@vger.kernel.org>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Macpaul Lin <macpaul.lin@mediatek.com>
CR-Id: ALPS05024007
Feature: Google Security Patch
(cherry-pick from 6959ee1fb88749e460ac2cc154719db17d3147c4)