[ALPS05118436]vt: selection, push console lock down
We need to nest the console lock in sel_lock, so we have to push it down
a bit. Fortunately, the callers of set_selection_* just lock the console
lock around the function call. So moving it down is easy.

In the next patch, we switch the order.

Bug : 148713787
Feature: Google Security Patch
CR-Id: ALPS05118436
Change-Id: Id2c59634d86490a5771bfbe7d724a4b2f7544c4d
Signed-off-by: mtk21421 <Stan.Lu@mediatek.com>
(cherry-pick from 827ef83bb3cacc8b0ce1c4fce419ef5c6e0cb486)